feat: update codestyle, change filenames, add moduleinfo.json, add doc
