+++ docs/install.md
@@ -1,0 +1,12 @@
+### EINE SICHERUNG DER DATENBANK ERSTELLEN
+###### Die Tabellen `configuration` und `admin_access` werden geändert.
+###### Es wird eine neue Tabelle `t10_searchstats` erstellt.
+
+---
+
+# Installation
+1. Loggen Sie sich in den Admin-Bereich ein
+2. Öffnen Sie die *Menüoption Module > Systemmodule*.
+3. Wählen Sie das **T10 Suchbegriffstatistik** Modul und klicken Sie auf Installieren..
+
+---
