+++ new_files/admin/includes/extra/csrf_exclusion/mcm_searchstats.php
@@ -1,0 +1,2 @@
+<?php
+$user_exclusions[] = 'mcm_searchstats.php';
