+++ new_files/admin/includes/extra/filenames/mcm_searchstats.php
@@ -1,2 +1,2 @@
 <?php
-define('FILENAME_MODULE_T10_SEARCHSTATS', 't10_searchstats.php');
+define('FILENAME_MODULE_T10_SEARCHSTATS', 'mcm_searchstats.php');
